integ: add a 'testshell' vte driver
This driver is necessary for testing bash because bash, even compared to
'bash -i', behaves differently when under a pty.

We need the bash behaviour of running PROMPT_COMMAND in order to test
pazi correctly, and I actually can't think of a better way to deal with
bash's garbage interactive mode other than this: creating a full pty and
parsing the vte output.

It's possible that there's a better solution out there, but honestly
this is attractive for other reasons too.
Benchmarking in this environment will be much more "real world".
Sure, the cost was my sanity since working with pty's is actual hell,
but I think what I've got here more or less works.