frecent_paths: dedupe entries by path
Also order things more consistently; it's always least-frecent ->
more-frecent in the paths coming out of frecent_paths, and thus main
should take 'last'.

This is primarily because '-i' and 'z' (with no args) order them as
most-important-last since the bottom of the list is more immediately
reachable for a viewer from the input prompt.