bench: turn of zle for zsh benchmarks
This makes tests more reliable by removing annoying behavior. It also
lets the testshell harness work better.

This change actually doubles pazi's benchmarked speed and makes it beat
out fasd.

It's incredibly suspicious that this makes so much of a difference, and
I'll still look into why with zle pazi's method of hooking zsh is so
*slow*, but I still think this is a right change to make for the sake of
the test framework.