bench/fasd: workaround fasd 'empty history' bug
Fasd, in some some bash configuratoins, fails in its prompt function if
the history file is empty. This is due to
https://github.com/clvv/fasd/blob/90b531a5daaa545c74c7d98974b54cbdb92659fc/fasd#L127-L130
implicitly assuming 'history 1' returns something.

Note, this bug appears to occur on bash 4.3+ubuntu but not 4.4+fedora.
As such, it happens on travis-ci even though it doesn't impact my local
environment.

It seems possible that travis's bash is setting 'shift_verbose' leading
to this, but frankly now that it works I'm going to pretend none of this
happened.