benches: add a cd bench
Due to pazi and fasd operating in significantly different ways, namely
on zsh using a hook vs parsing prompt input, and on bash vs checking
only prompt PWD vs parsing all args, this test is probably a more
interesting one.

People typically do not do highly '&&'d commands. However, in the cd_50
bench, pazi's method of working is invoked once per '&&' while fasd is
invoked once for the whole thing.

This difference does matter, so I'm leaving the cd_50 in, but this is a
more useful comparison for regular use.