tests/bench: allow 'wait'ing for forking jumpers
This implements a way for benchmarks to indicate that the testshell
should wait for any forked processes to end.

This is helpful for `z` and `autojump`, which add directories by forking
off into the background.

This will allow benchmarking the speed of those operations, and to make
the benchmarks not flaky (e.g. rcd + jump can flake without a wait in
between).

The implementation is done by using cgroups to keep track of pids,
assuming a shell during normal operation will only have one pid, and
then going from there.

There are a few significant caveats:
1. cgroups are created with systemd-run scopes; this requires root
   pending a bug on their end.
2. This requires cgroups-v2 or hybrid for tracking processes
   (cgroup.procs) vs threads.
3. The overhead introduced by this is difficult to measure perfectly; it
   has no significant impact in the case where children aren't forked,
   but when children are, I have no baseline to compare with.

This commit also does a bit of other cleanup, such as switching the
harness to a builder method.