main: move 'jump' and 'visit' to subcommands
This fixes #70.

Prior to this change, there was a mix of subcommands and flags that were
all mutually-exclusive behaviors.
This ended up being a bit awkward and made some configurations
ambiguous.

This moves to a much nicer structure. There are global flags (right now
just debug), and then there are subcommands for each distinct
mode of operation.

For ease of migration, the previous logic still remains, but it can be
safely deleted after another release or so to allow migration.