tests: add support for fish to integ tests
This removes the conch-shell joke in favor of a working fish shell.

This also makes various minor tweaks required for fish to work,
including the following:

1. Support multi-byte characters in our VTE line management.
Fish prints a multi-byte '⏎' char on exit. This would cause a panic
because '.truncate(1)' would hit halfway through the character. We now
track multi-byte characters correctly too.
2. Hack around fish-greeting and fish-completions.
3. Begin splitting out pieces of commands we run into shell-specific
functions to account for fish differences.
4. Sets 'TERM=xterm' because fish cares.