frecency: Boost interactive selection frecency
This gives a 3x weighted boost to the result of an interactive selection
on the assumption that if the user went out of their way to do an
interactive selection to choose something it's something they wish to
visit more often.

This does the necessary plumbing to implement the feature, and
implements it.