tests: wait in all tests; add pgrep waiter
Now that we disown in pazi, we always need to wait.

However, we also want to test on macos and non-root-linux, where
cgroups are not a viable way to track children.

This uses `pgrep -P` to track children, which works on all platforms we
care about, and should work just as well (other than taking more
processes to do so in theory being slower).