tests: temporarily disable fish tests
They fail on macOS, and on linux with a new enough fish version.

The reason isn't totally clear to me. Something about fish's output has
changed in a way that breaks the vte/harness code, and I don't have time
to debug it now.

Turn them off for now, we'll come back to this later.